requirements.txt
fastapi[standard]
uvicorn
python-jose
bcrypt
python-dotenv
pydantic
pydantic-settings
aiofiles
motor